requirements.txt
django >= 2.1.6, < 3.0 
requests >= 2.22.0
beautifulsoup4 >= 4.8.2
shapely >= 1.7
pyshp >= 2.1.0
networkx >= 2.4
docutils >= 0.16
djangorestframework >= 3.11.0
django-filter >= 2.2.0
django-cors-headers >= 3.2.1